Fix small errors
