+++ .gitignore
@@ -1,4 +1,5 @@
 *.js
+!tests/*.js
 *.DS_Store
 *.log
 npm-debug.log*
