+++ tslint.json
@@ -60,6 +60,7 @@
         "no-any": false,
         "no-arg": true,
         "no-conditional-assignment": true,
+        "no-consecutive-blank-lines": true,
         "no-construct": true,
         "no-debugger": true,
         "no-default-export": true,
